Add files via upload
Simulation of the reaction-diffusion-advection equation with the TVD stabilisation of the convective term.
The current setting is the "solid body rotation" benchmark:  simulation of the pure transport equation on the unit square [0,1]^2, where initial condition transports in the counter-clockwise direction. 
The code is based on the dealii finite element library. 
To let it run, one has to install dealii and then include these source *.cc files into the project (see dealii/examples/, e.g. the example 12 of a heat equation).
11:11     11.11.2016